+++ htdocs/components/customerEditMain/CustomerNotesComponent.js
@@ -1,342 +1,292 @@
-import React, {Fragment} from 'react';
+import moment from "moment";
+import React, { Fragment } from "react";
 import Skeleton from "react-loading-skeleton";
+import APICustomers from "../services/APICustomers";
+import APIUser from "../services/APIUser";
+import MainComponent from "../shared/MainComponent";
+import Modal from "../shared/Modal/modal";
+import Spinner from "../shared/Spinner/Spinner";
+import ToolTip from "../shared/ToolTip";
+import Table from "./../shared/table/table";
 
-class CustomerNotesComponent extends React.PureComponent {
-    el = React.createElement;
+class CustomerNotesComponent extends MainComponent {
+  el = React.createElement;
+  api = new APICustomers();
+  apiUsers = new APIUser();
 
-    constructor(props) {
-        super(props);
-        this.state = {
-            loaded: false,
-            customerNotes: [],
-            customerId: props.customerId,
-            currentNote: null,
-            currentNoteIdx: null,
-            isAddingNote: false
-        };
-        document.customerNotesComponent = this;
-        this.handleNoteTextChange = this.handleNoteTextChange.bind(this);
-    }
+  constructor(props) {
+    super(props);
+    this.state = {
+      ...this.state,
+      loaded: false,
+      showModal:false,
+      customerNotes: [],
+      enginners: [],
+      customerId: props.customerId,
+      currentNote: null,
+      currentNoteIdx: null,
+      isAddingNote: false,
+      reviewData: {
+        toBeReviewedOnAction: "",
+        toBeReviewedOnByEngineerId: "",
+        toBeReviewedOnDate: "",
+        toBeReviewedOnTime: "",
+      },
+    };
+    
+   }
 
-    fetchCustomerNotes() {
-        return fetch('/CustomerNote.php?action=getCustomerNotes&customerId=' + this.props.customerId)
-            .then(response => response.json())
-            .then(response => this.setState({customerNotes: response.data}));
-    }
+  componentDidMount() {
+    this.api.getCustomerReviewData(this.state.customerId).then(
+      (res) => {
+        console.log("reviewData", res);
+        this.setState({ reviewData: res.data });
+      },
+      (error) => {
+        this.alert("Error in loading data");
+      }
+    );
+    this.apiUsers.getActiveUsers().then((enginners) => {
+      this.setState({ enginners });
+    });
+    this.getData();
+  }
+  getData = () => {
+    this.api.getCustomerNotes(this.props.customerId).then((response) => {
+        console.log('notes',response)
+      this.setState({ customerNotes: response.data ,loaded:true,showModal:false});
+    });
+  };
+ 
+  
+ 
+  handleEdit = (note) => {
+    console.log(note);
+    this.setState({ currentNote: {...note},showModal:true });
+  };
+  getHistoryElement() {
+    const { customerNotes } = this.state;
+    const columns = [
+      {
+        path: "modifiedAt",
+        label: "",
+        hdToolTip: "Modefied At",
+        icon: "fal fa-2x fa-calendar color-gray2 pointer",
+        sortable: true,
+        width: 150,
+        content: (note) => this.getCorrectDate(note.modifiedAt, true),
+      },
+      {
+        path: "modifiedByName",
+        label: "",
+        hdToolTip: "Modefied By",
+        icon: "fal fa-2x fa-user-hard-hat color-gray2 pointer",
+        sortable: true,
+        width: 120,
+      },
+      {
+        path: "note",
+        label: "",
+        hdToolTip: "Note",
+        icon: "fal fa-2x fa-file color-gray2 pointer",
+        sortable: true,
+      },
+      {
+        path: "",
+        label: "",
+        hdToolTip: "Edit",
+        icon: this.getEditIcon(),
+        sortable: true,
+        content: (note) => this.getEditElement(note, this.handleEdit),
+      },
+      {
+        path: "delete",
+        label: "",
+        hdToolTip: "Delete",
+        icon: this.getDeleteElement(),
+        sortable: true,
+        content: (note) =>
+          this.getDeleteElement(note, this.handleDeleteNote),
+      },
+    ];
+    return (
+      <Table
+        pk="id"
+        data={customerNotes}
+        columns={columns}
+        search={true}
+      ></Table>
+    );
+  }
+  handleDeleteNote =async (note) => {
+      if(await this.confirm("Are you sure you want to detele this note?"))
+      {
+        this.api.deleteCustomerNote(note.id).then((res) => this.getData());
+      }
+  };
+  
+  
+  handleSaveNote = () => {
+    const { currentNote ,customerId} = this.state;
+    if (this.state.currentNote.id >= 0)
+      currentNote.lastUpdatedDateTime = moment().format("YYYY-MM-DD HH:mm:ss");
+    currentNote.customerId=customerId;
+    this.api
+      .saveCustomerNote(currentNote)
+      .then((response) => {
+        this.getData();
+      })
+      .catch((error) => {
+        console.log(error);
+        this.alert("Failed to save note");
+      });
+  };
+  
+ 
+  setReviewData = ($event) => {
+    const { reviewData } = this.state;
+    reviewData[$event.target.name] = $event.target.value;
+    this.setState({ reviewData });
+  };
+  handleSaveREviewData = () => {
+    const { reviewData, customerId } = this.state;
+    reviewData.customerId = customerId;
+    this.api.updateCustomerReviewData(reviewData).then(
+      (res) => {
+        this.alert("saved");
+      },
+      (error) => {
+        this.alert("Error in save data");
+      }
+    );
+  };
+  getReviewElement = () => {
+    const { reviewData } = this.state;
+    return (
+      <div className="flex-row" style={{ alignItems: "center" }}>
+        <span>To be reviewed on</span>
+        <input
+          style={{ width: 140 }}
+          className="form-control"
+          type="date"
+          value={reviewData.toBeReviewedOnDate}
+          name="toBeReviewedOnDate"
+          onChange={($event) => this.setReviewData($event)}
+        ></input>
+        <span>Time</span>
+        <input
+          style={{ width: 80 }}
+          className="form-control"
+          type="time"
+          value={reviewData.toBeReviewedOnTime}
+          name="toBeReviewedOnTime"
+          onChange={($event) => this.setReviewData($event)}
+        ></input>
+        <span>By</span>
+        <select
+          style={{ width: 150 }}
+          className="form-control"
+          value={reviewData.toBeReviewedOnByEngineerId}
+          name="toBeReviewedOnByEngineerId"
+          onChange={($event) => this.setReviewData($event)}
+        >
+          {this.state.enginners.map((e) => (
+            <option key={e.id} value={e.id}>
+              {e.name}
+            </option>
+          ))}
+        </select>
+        <ToolTip title="Save reviewed data">
+          <i
+            onClick={this.handleSaveREviewData}
+            className="fal fa-save fa-2x icon pointer "
+          ></i>
+        </ToolTip>
+      </div>
+    );
+  };
+  setNote=(value)=>{
+    const {currentNote } = this.state;
+    currentNote.note=value;
+    this.setState({currentNote})
+  }
+  getModal=()=>{
+      const { showModal,currentNote } = this.state;
+      if(!showModal)
+      return null;
+      return (
+        <Modal
+          width={600}
+          show={showModal}
+          title="Update Customer Note"
+          footer={
+            <div key="footerActions">
+              <button onClick={this.handleSaveNote}>Save</button>
+              <button onClick={() => this.setState({ showModal: false })}>
+                Cancel
+              </button>
+            </div>
+          }
+        >
+          <div>
+              {
+                  currentNote.id>=0?
+              <div>
+                  <span>Updated by : </span>
+                  <span>{currentNote.modifiedByName}</span>
+                  <span>   At </span>
+                  <span>{this.getCorrectDate(currentNote.modifiedAt,true)}</span>
+              </div>:null
+            }
+            <textarea
+            value={currentNote.note}
+            onChange={($event)=>this.setNote($event.target.value)}
+              className="form-control"
+              style={{ height: 200 }}
+            ></textarea>
+          </div>
+        </Modal>
+      );
+  }
+ 
+  handleNewNote = () => {
+    this.setState({
+      currentNote: {
+        id: -1,
+        customerId: this.props.customerId,
+        createdAt: "",
+        modifiedAt: "",
+        modifiedById: "",
+        note: "",
+        createdById: "",
+        modifiedByName: "",
+      },
+      showModal:true
+    });
+  };
+  getAddNewNoteElement=()=>{
+      return <div className="m-5"> <ToolTip width={30} title="Add New Note">
+          <i className="fal fa-plus fa-2x pointer" onClick={this.handleNewNote}></i>
+      </ToolTip>
+      </div>
+  }
+  render() {
+    if (!this.state.loaded) {
+      return <Spinner show={this.state.loaded}></Spinner>;
+    }   
 
-    componentDidMount() {
-        this.fetchCustomerNotes()
-            .then(() => {
-                this.setState({
-                    loaded: true,
-                    currentNote: this.state.customerNotes.length ? this.state.customerNotes[0] : null,
-                    currentNoteIdx: 0,
-                });
-            });
-    }
-
-    getIcon(iconClass) {
-        return this.el('i', {className: iconClass});
-    }
-
-    createButton(className, iconClass, clickFunction, text) {
-        return (
-            'button',
-                {
-                    key: className,
-                    className,
-                    type: 'button',
-                    onClick: clickFunction
-                },
-                [
-                    this.getIcon(iconClass),
-                    text
-                ]
-        )
-    }
-
-    changeNote(nextIdx) {
-        if (nextIdx >= 0 || nextIdx <= this.state.customerNotes.length - 1) {
-            this.setState({
-                currentNote: {...this.state.customerNotes[nextIdx]},
-                currentNoteIdx: nextIdx,
-                isAddingNote: false
-            });
-        } else {
-            this.setState(
-                {
-                    currentNote: null,
-                    currentNoteIdx: 0,
-                    isAddingNote: false
-                }
-            )
-        }
-
-    }
-
-
-    renderNotes() {
-
-        <Fragment>
-
-        </Fragment>
-
-        return this.el(
-            React.Fragment,
-            {
-                key: 'notesHolder'
-            },
-            this.el(
-                'div',
-                {
-                    style: {
-                        display: 'flex',
-                        flexDirection: 'column',
-                        height: "400px",
-                        overflowY: 'scroll',
-                    }
-                },
-                [
-                    ...this.state.customerNotes.map(x => {
-                        return this.el(
-                            'div',
-                            {
-                                key: `rowsHolder-${x.id}`,
-                            },
-                            [
-                                this.el(
-                                    'div',
-                                    {
-                                        key: `basicInfoRow-${x.id}`,
-                                        style: {
-                                            backgroundColor: '#cccccc'
-                                        }
-                                    },
-                                    this.el(
-                                        'span',
-                                        {
-                                            key: `date-${x.id}`,
-
-                                        },
-                                        [
-                                            x.modifiedAt,
-                                            ' - ',
-                                            x.modifiedByName
-                                        ]
-                                    )
-                                ),
-                                this.el(
-                                    'div',
-                                    {
-                                        key: `detailsRow-${x.id}`
-                                    },
-                                    x.note
-                                )
-                            ]
-                        )
-                    })
-                ]
-            )
-        )
-    }
-
-    handleNoteTextChange($event) {
-        this.setState(
-            {
-                currentNote: {...this.state.currentNote, note: $event.target.value}
-            });
-    }
-
-    render() {
-
-        if (!this.state.loaded) {
-            return this.el(
-                Skeleton,
-                null,
-                'Loading Data'
-            );
-        }
-
-
-        return (
-            <Fragment>
-                <div className="form-group customerNoteHistory">
-                    {this.renderNotes()}
-                    <div className="customerNoteNav mt-3 mb-3">
-                        <button type="button"
-                                name="First"
-                                aria-hidden="true"
-                                onClick={() => {
-                                    const nextIdx = 0;
-                                    this.changeNote(nextIdx);
-                                }}
-                                className="btn btn-outline-secondary"
-                        >
-                            <i className="fa fa-step-backward">
-                            </i> First
-                        </button>
-                        <button type="button"
-                                name="Previous"
-                                onClick={
-                                    () => {
-                                        const nextIdx = this.state.currentNoteIdx - 1;
-                                        this.changeNote(nextIdx);
-                                    }
-                                }
-                                className="btn btn-outline-secondary"
-                        >
-                            <i className="fa fa-backward"
-                               aria-hidden="true"
-                            >
-                            </i> Back
-                        </button>
-                        <button type="button"
-                                name="Next"
-                                onClick={
-                                    () => {
-                                        const nextIdx = this.state.currentNoteIdx + 1;
-                                        this.changeNote(nextIdx);
-                                    }
-                                }
-                                className="btn btn-outline-secondary"
-                        >
-                            Next
-                            <i className="fa fa-forward"
-                               aria-hidden="true"
-                            />
-                        </button>
-
-                        <button type="button"
-                                name="Last"
-                                onClick={
-                                    () => {
-                                        const nextIdx = this.state.customerNotes.length - 1;
-                                        this.changeNote(nextIdx);
-                                    }
-                                }
-                                className="btn btn-outline-secondary"
-                        >
-                            Last
-                            <i className="fa fa-step-forward"
-                               aria-hidden="true"
-                            />
-                        </button>
-                        {
-
-                        }
-                        <button type="button"
-                                name="Delete"
-                                onClick={
-                                    () => {
-                                        if (!this.state.currentNote) {
-                                            return;
-                                        }
-                                        const currentNoteId = this.state.currentNote.id;
-
-                                        fetch(`/CustomerNote.php?action=deleteNote&noteId=${currentNoteId}`)
-                                            .then(() => {
-                                                this.setState({
-                                                    customerNotes: [...this.state.customerNotes.filter(x => x.id !== currentNoteId)]
-                                                });
-                                                const nextIdx = 0;
-                                                this.changeNote(nextIdx);
-                                            })
-                                    }
-                                }
-                                className="btn btn-outline-danger"
-                        >
-                            <i className="fa fa-trash"
-                               aria-hidden="true"
-                            />
-                            Delete
-                        </button>
-                        <button type="button"
-                                name="New"
-                                onClick={
-                                    () => {
-                                        this.setState({
-                                            currentNote: {
-                                                id: -1,
-                                                customerId: this.props.customerId,
-                                                createdAt: '',
-                                                modifiedAt: '',
-                                                modifiedById: '',
-                                                note: '',
-                                                createdById: '',
-                                                modifiedByName: '',
-                                            },
-                                            isAddingNote: true,
-                                        })
-                                    }
-                                }
-                                className="btn btn-outline-secondary"
-                        >
-                            <i className="fa fa-plus-circle"
-                               aria-hidden="true"
-                            />
-                            New
-                        </button>
-                        <button type="button"
-                                name="Save"
-                                onClick={
-                                    () => {
-                                        fetch('/CustomerNote.php?action=updateNote',
-                                            {
-                                                method: 'POST',
-                                                headers: {
-                                                    dataType: 'application/json'
-                                                },
-                                                body: JSON.stringify(this.state.currentNote)
-                                            }
-                                        )
-                                            .then(response => response.json())
-                                            .then(response => {
-                                                if (this.state.isAddingNote) {
-                                                    this.setState(
-                                                        {
-                                                            customerNotes: [response.data, ...this.state.customerNotes],
-                                                            currentNote: {...response.data},
-                                                            currentNoteIdx: 0,
-                                                            isAddingNote: false
-                                                        }
-                                                    )
-                                                } else {
-                                                    this.setState({
-                                                            customerNotes: [...this.state.customerNotes.map(x => x.id === response.data.id ? response.data : x)],
-                                                            currentNote: {...response.data}
-                                                        }
-                                                    )
-                                                }
-                                            })
-                                            .catch(error => {
-                                                alert('Failed to save note');
-                                            })
-                                    }
-                                }
-                                className="btn btn-outline-secondary"
-                        >
-                            <i className="fa fa-floppy-o"
-                               aria-hidden="true"
-                            />
-
-                            Save
-                        </button>
-                    </div>
-                </div>
-                <div className="form-group customerNoteDetails">
-                    <textarea name="customerNoteDetails"
-                              id="customerNoteDetails"
-                              cols="120"
-                              value={this.state.currentNote && this.state.currentNote.note || ''}
-                              rows="12"
-                              className="form-control"
-                              onChange={($event) => this.handleNoteTextChange($event)}
-                    />
-                </div>
-            </Fragment>
-        )
-
-    }
+    return (
+      <div>
+        {this.getConfirm()}
+        {this.getAlert()}
+        {this.getModal()}        
+        {this.getReviewElement()}
+        {this.getAddNewNoteElement()}
+        {this.getHistoryElement()}
+     
+      </div>
+    );
+  }
 }
 
 export default CustomerNotesComponent;